fix(task-graph): defer the pipe-function wrapper past the Task cycle
`context.own(fn)` threw "Class extends value undefined" for every plain
function. Task.ts imports TaskRunner, which imports Conversions, so
Conversions evaluates while Task is still uninitialized (confirmed by
probe: `typeof Task` is undefined at Conversions' module scope). A
method body that merely reads Task later is fine -- ensureTask's
`instanceof Task` works, and `typeof Task` is "function" by call time --
but a `class ... extends Task` clause resolves the superclass against
the binding as it stood at evaluation, so it saw undefined.

Conversions already documents this exact cycle and solves it for the
GraphAsTask wrapper with a deferred factory seam; the pipe-function
wrapper was simply missed. This applies the same idiom: PipeFunctionTask
owns the subclass and registers the factory once Task exists. If that
registration ordering ever breaks, the seam throws a named error rather
than the cryptic extends failure.

Also fixes source mode skipping private workspaces. findWorkspaces()
filtered to publishConfig.access === "public", which is right for
publishing but wrong for stubbing: providers/aws and providers/cloudflare
are private, so use-source never wrote their dist stubs and 11 test files
could not resolve @workglow/aws/job-queue locally. They passed in CI,
which builds for real, so this was local-only red that trains people to
ignore red. The filter is now a parameter, defaulting to the previous
behavior for publish and link.
